[move] SecurityConfig 위치를 global/config로 이동, CustomUserDetailService 위치를 global/auth로 이동
