+++ src/main/java/chungbazi/chungbazi_be/global/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package chungbazi.chungbazi_be.config;
+package chungbazi.chungbazi_be.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
